tripletex/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
requests==2.32.0
anthropic==0.43.0
python-dotenv==1.0.1
python-multipart==0.0.9
